# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * سَنَد - المنظومة الصحية الوطنية API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface Patient {
  id: number;
  nationalId: string;
  nameAr: string;
  nameEn: string;
  bloodType: string;
  dateOfBirth: string;
  gender: string;
  city: string;
  phone: string;
}

export interface EmergencyContact {
  name: string;
  relation: string;
  phone: string;
}

export interface Medication {
  name: string;
  nameAr: string;
  dosage: string;
  frequency: string;
  prescribedBy: string;
  hospital: string;
  prescribedDate: string;
}

export interface MedicalRecord {
  id: number;
  date: string;
  hospital: string;
  hospitalAr: string;
  doctorName: string;
  specialty: string;
  specialtyAr: string;
  diagnosis: string;
  diagnosisAr: string;
  notes: string;
  type: string;
}

export interface LabResult {
  id: number;
  testName: string;
  testNameAr: string;
  date: string;
  hospital: string;
  value: string;
  unit: string;
  normalRange: string;
  status: string;
}

export interface Vaccination {
  name: string;
  nameAr: string;
  date: string;
  nextDue: string;
  status: string;
}

export interface PatientFull {
  id: number;
  nationalId: string;
  nameAr: string;
  nameEn: string;
  bloodType: string;
  dateOfBirth: string;
  gender: string;
  city: string;
  phone: string;
  allergies: string[];
  chronicConditions: string[];
  emergencyContacts: EmergencyContact[];
  currentMedications: Medication[];
  medicalRecords: MedicalRecord[];
  labResults: LabResult[];
  vaccinations: Vaccination[];
}

export interface EmergencyData {
  nationalId: string;
  nameAr: string;
  bloodType: string;
  allergies: string[];
  chronicConditions: string[];
  currentMedications: string[];
  emergencyContacts: EmergencyContact[];
  criticalNotes: string;
}

export type PhysicianDashboardSummary = {
  totalVisits: number;
  hospitalsVisited: number;
  activeMedications: number;
  lastVisit: string;
  riskScore: string;
};

export interface DrugAlert {
  severity: string;
  drug1: string;
  drug2: string;
  message: string;
  messageAr: string;
}

export interface PhysicianDashboard {
  patient: PatientFull;
  summary: PhysicianDashboardSummary;
  drugInteractionAlerts: DrugAlert[];
}

export interface DrugInteractionRequest {
  newDrug: string;
  existingDrugs: string[];
}

export interface DrugInteractionResult {
  hasInteraction: boolean;
  severity: string;
  alerts: DrugAlert[];
}

export interface AiPrediction {
  condition: string;
  conditionAr: string;
  probability: number;
  timeframe: string;
  severity: string;
  actionRequired: string;
  actionRequiredAr: string;
}

export interface AiPredictions {
  nationalId: string;
  riskLevel: string;
  predictions: AiPrediction[];
  recommendations: string[];
}

export interface CitizenDashboard {
  patient: Patient;
  healthScore: number;
  recentVisits: MedicalRecord[];
  upcomingVaccinations: Vaccination[];
  currentMedications: Medication[];
  aiAlerts: AiPrediction[];
  labResults: LabResult[];
}

export interface SystemStats {
  totalPatients: number;
  totalHospitals: number;
  totalMedicalRecords: number;
  drugInteractionsPrevented: number;
  emergencyResponsesThisMonth: number;
  activePhysicians: number;
}