# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: سَنَد - المنظومة الصحية الوطنية API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health check
  - name: patients
    description: Patient management
  - name: emergency
    description: Emergency mode - paramedic interface
  - name: physician
    description: Physician dashboard
  - name: medications
    description: Medications and drug interactions
  - name: ai
    description: AI predictions and alerts
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"

  /patients/{nationalId}:
    get:
      operationId: getPatientByNationalId
      tags: [patients]
      summary: Get patient by national ID
      parameters:
        - name: nationalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Patient record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientFull"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /emergency/{nationalId}:
    get:
      operationId: getEmergencyData
      tags: [emergency]
      summary: Get emergency critical data for a patient (paramedic mode)
      parameters:
        - name: nationalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Emergency critical data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyData"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /physician/{nationalId}:
    get:
      operationId: getPhysicianDashboard
      tags: [physician]
      summary: Get full medical history for physician dashboard
      parameters:
        - name: nationalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full patient medical history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhysicianDashboard"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /medications/check-interaction:
    post:
      operationId: checkDrugInteraction
      tags: [medications]
      summary: Check for drug interactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrugInteractionRequest"
      responses:
        "200":
          description: Drug interaction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrugInteractionResult"

  /ai/predictions/{nationalId}:
    get:
      operationId: getAiPredictions
      tags: [ai]
      summary: Get AI health predictions for a patient
      parameters:
        - name: nationalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI health predictions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiPredictions"

  /citizen/{nationalId}:
    get:
      operationId: getCitizenDashboard
      tags: [patients]
      summary: Get citizen health dashboard
      parameters:
        - name: nationalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Citizen health dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CitizenDashboard"

  /stats:
    get:
      operationId: getSystemStats
      tags: [health]
      summary: Get system-wide statistics
      responses:
        "200":
          description: System statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    Patient:
      type: object
      properties:
        id:
          type: integer
        nationalId:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        bloodType:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        city:
          type: string
        phone:
          type: string
      required:
        - id
        - nationalId
        - nameAr
        - nameEn
        - bloodType
        - dateOfBirth
        - gender
        - city
        - phone

    PatientFull:
      type: object
      properties:
        id:
          type: integer
        nationalId:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        bloodType:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        city:
          type: string
        phone:
          type: string
        allergies:
          type: array
          items:
            type: string
        chronicConditions:
          type: array
          items:
            type: string
        emergencyContacts:
          type: array
          items:
            $ref: "#/components/schemas/EmergencyContact"
        currentMedications:
          type: array
          items:
            $ref: "#/components/schemas/Medication"
        medicalRecords:
          type: array
          items:
            $ref: "#/components/schemas/MedicalRecord"
        labResults:
          type: array
          items:
            $ref: "#/components/schemas/LabResult"
        vaccinations:
          type: array
          items:
            $ref: "#/components/schemas/Vaccination"
      required:
        - id
        - nationalId
        - nameAr
        - nameEn
        - bloodType
        - dateOfBirth
        - gender
        - city
        - phone
        - allergies
        - chronicConditions
        - emergencyContacts
        - currentMedications
        - medicalRecords
        - labResults
        - vaccinations

    EmergencyContact:
      type: object
      properties:
        name:
          type: string
        relation:
          type: string
        phone:
          type: string
      required:
        - name
        - relation
        - phone

    Medication:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        prescribedBy:
          type: string
        hospital:
          type: string
        prescribedDate:
          type: string
      required:
        - name
        - nameAr
        - dosage
        - frequency
        - prescribedBy
        - hospital
        - prescribedDate

    MedicalRecord:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        hospital:
          type: string
        hospitalAr:
          type: string
        doctorName:
          type: string
        specialty:
          type: string
        specialtyAr:
          type: string
        diagnosis:
          type: string
        diagnosisAr:
          type: string
        notes:
          type: string
        type:
          type: string
      required:
        - id
        - date
        - hospital
        - hospitalAr
        - doctorName
        - specialty
        - specialtyAr
        - diagnosis
        - diagnosisAr
        - notes
        - type

    LabResult:
      type: object
      properties:
        id:
          type: integer
        testName:
          type: string
        testNameAr:
          type: string
        date:
          type: string
        hospital:
          type: string
        value:
          type: string
        unit:
          type: string
        normalRange:
          type: string
        status:
          type: string
      required:
        - id
        - testName
        - testNameAr
        - date
        - hospital
        - value
        - unit
        - normalRange
        - status

    Vaccination:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        date:
          type: string
        nextDue:
          type: string
        status:
          type: string
      required:
        - name
        - nameAr
        - date
        - nextDue
        - status

    EmergencyData:
      type: object
      properties:
        nationalId:
          type: string
        nameAr:
          type: string
        bloodType:
          type: string
        allergies:
          type: array
          items:
            type: string
        chronicConditions:
          type: array
          items:
            type: string
        currentMedications:
          type: array
          items:
            type: string
        emergencyContacts:
          type: array
          items:
            $ref: "#/components/schemas/EmergencyContact"
        criticalNotes:
          type: string
      required:
        - nationalId
        - nameAr
        - bloodType
        - allergies
        - chronicConditions
        - currentMedications
        - emergencyContacts
        - criticalNotes

    PhysicianDashboard:
      type: object
      properties:
        patient:
          $ref: "#/components/schemas/PatientFull"
        summary:
          type: object
          properties:
            totalVisits:
              type: integer
            hospitalsVisited:
              type: integer
            activeMedications:
              type: integer
            lastVisit:
              type: string
            riskScore:
              type: string
          required:
            - totalVisits
            - hospitalsVisited
            - activeMedications
            - lastVisit
            - riskScore
        drugInteractionAlerts:
          type: array
          items:
            $ref: "#/components/schemas/DrugAlert"
      required:
        - patient
        - summary
        - drugInteractionAlerts

    DrugAlert:
      type: object
      properties:
        severity:
          type: string
        drug1:
          type: string
        drug2:
          type: string
        message:
          type: string
        messageAr:
          type: string
      required:
        - severity
        - drug1
        - drug2
        - message
        - messageAr

    DrugInteractionRequest:
      type: object
      properties:
        newDrug:
          type: string
        existingDrugs:
          type: array
          items:
            type: string
      required:
        - newDrug
        - existingDrugs

    DrugInteractionResult:
      type: object
      properties:
        hasInteraction:
          type: boolean
        severity:
          type: string
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/DrugAlert"
      required:
        - hasInteraction
        - severity
        - alerts

    AiPredictions:
      type: object
      properties:
        nationalId:
          type: string
        riskLevel:
          type: string
        predictions:
          type: array
          items:
            $ref: "#/components/schemas/AiPrediction"
        recommendations:
          type: array
          items:
            type: string
      required:
        - nationalId
        - riskLevel
        - predictions
        - recommendations

    AiPrediction:
      type: object
      properties:
        condition:
          type: string
        conditionAr:
          type: string
        probability:
          type: number
        timeframe:
          type: string
        severity:
          type: string
        actionRequired:
          type: string
        actionRequiredAr:
          type: string
      required:
        - condition
        - conditionAr
        - probability
        - timeframe
        - severity
        - actionRequired
        - actionRequiredAr

    CitizenDashboard:
      type: object
      properties:
        patient:
          $ref: "#/components/schemas/Patient"
        healthScore:
          type: integer
        recentVisits:
          type: array
          items:
            $ref: "#/components/schemas/MedicalRecord"
        upcomingVaccinations:
          type: array
          items:
            $ref: "#/components/schemas/Vaccination"
        currentMedications:
          type: array
          items:
            $ref: "#/components/schemas/Medication"
        aiAlerts:
          type: array
          items:
            $ref: "#/components/schemas/AiPrediction"
        labResults:
          type: array
          items:
            $ref: "#/components/schemas/LabResult"
      required:
        - patient
        - healthScore
        - recentVisits
        - upcomingVaccinations
        - currentMedications
        - aiAlerts
        - labResults

    SystemStats:
      type: object
      properties:
        totalPatients:
          type: integer
        totalHospitals:
          type: integer
        totalMedicalRecords:
          type: integer
        drugInteractionsPrevented:
          type: integer
        emergencyResponsesThisMonth:
          type: integer
        activePhysicians:
          type: integer
      required:
        - totalPatients
        - totalHospitals
        - totalMedicalRecords
        - drugInteractionsPrevented
        - emergencyResponsesThisMonth
        - activePhysicians