# lib/api-zod/src/generated/types/aiPrediction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * سَنَد - المنظومة الصحية الوطنية API
 * OpenAPI spec version: 0.1.0
 */

export interface AiPrediction {
  condition: string;
  conditionAr: string;
  probability: number;
  timeframe: string;
  severity: string;
  actionRequired: string;
  actionRequiredAr: string;
}