# lib/api-zod/src/generated/types/aiPredictions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * سَنَد - المنظومة الصحية الوطنية API
 * OpenAPI spec version: 0.1.0
 */
import type { AiPrediction } from "./aiPrediction";

export interface AiPredictions {
  nationalId: string;
  riskLevel: string;
  predictions: AiPrediction[];
  recommendations: string[];
}