# lib/api-zod/src/generated/types/drugAlert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * سَنَد - المنظومة الصحية الوطنية API
 * OpenAPI spec version: 0.1.0
 */

export interface DrugAlert {
  severity: string;
  drug1: string;
  drug2: string;
  message: string;
  messageAr: string;
}