# lib/api-zod/src/generated/types/drugInteractionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * سَنَد - المنظومة الصحية الوطنية API
 * OpenAPI spec version: 0.1.0
 */
import type { DrugAlert } from "./drugAlert";

export interface DrugInteractionResult {
  hasInteraction: boolean;
  severity: string;
  alerts: DrugAlert[];
}