# lib/api-zod/src/generated/types/emergencyData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * سَنَد - المنظومة الصحية الوطنية API
 * OpenAPI spec version: 0.1.0
 */
import type { EmergencyContact } from "./emergencyContact";

export interface EmergencyData {
  nationalId: string;
  nameAr: string;
  bloodType: string;
  allergies: string[];
  chronicConditions: string[];
  currentMedications: string[];
  emergencyContacts: EmergencyContact[];
  criticalNotes: string;
}