# lib/api-zod/src/generated/types/labResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * سَنَد - المنظومة الصحية الوطنية API
 * OpenAPI spec version: 0.1.0
 */

export interface LabResult {
  id: number;
  testName: string;
  testNameAr: string;
  date: string;
  hospital: string;
  value: string;
  unit: string;
  normalRange: string;
  status: string;
}