# lib/api-zod/src/generated/types/medicalRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * سَنَد - المنظومة الصحية الوطنية API
 * OpenAPI spec version: 0.1.0
 */

export interface MedicalRecord {
  id: number;
  date: string;
  hospital: string;
  hospitalAr: string;
  doctorName: string;
  specialty: string;
  specialtyAr: string;
  diagnosis: string;
  diagnosisAr: string;
  notes: string;
  type: string;
}