# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * سَنَد - المنظومة الصحية الوطنية API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  nationalId: string;
  nameAr: string;
  nameEn: string;
  bloodType: string;
  dateOfBirth: string;
  gender: string;
  city: string;
  phone: string;
}