# lib/api-zod/src/generated/types/patientFull.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * سَنَد - المنظومة الصحية الوطنية API
 * OpenAPI spec version: 0.1.0
 */
import type { EmergencyContact } from "./emergencyContact";
import type { LabResult } from "./labResult";
import type { MedicalRecord } from "./medicalRecord";
import type { Medication } from "./medication";
import type { Vaccination } from "./vaccination";

export interface PatientFull {
  id: number;
  nationalId: string;
  nameAr: string;
  nameEn: string;
  bloodType: string;
  dateOfBirth: string;
  gender: string;
  city: string;
  phone: string;
  allergies: string[];
  chronicConditions: string[];
  emergencyContacts: EmergencyContact[];
  currentMedications: Medication[];
  medicalRecords: MedicalRecord[];
  labResults: LabResult[];
  vaccinations: Vaccination[];
}