# lib/api-zod/src/generated/types/physicianDashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * سَنَد - المنظومة الصحية الوطنية API
 * OpenAPI spec version: 0.1.0
 */
import type { DrugAlert } from "./drugAlert";
import type { PatientFull } from "./patientFull";
import type { PhysicianDashboardSummary } from "./physicianDashboardSummary";

export interface PhysicianDashboard {
  patient: PatientFull;
  summary: PhysicianDashboardSummary;
  drugInteractionAlerts: DrugAlert[];
}